clear out IDE managed auth data prior to git operation
